Create UpdatedSMILEPracticeScheduling_3.js
This connected the check to see if an email was scheduled and whether the email was sent. If the email had not been sent and it was past the day and time of the scheduled email, then this would activate the excel to send an email. What is not included is a trigger with google. The trigger used the function "checkemailstatus" and runs every minute to see if an email should be sent.